Sweep dead exports and split the two biggest files
One `sub` for the whole library. defs-mobile, defs-blocks-app,
defs-templates and defs-blocks-marketing each carried an identical copy,
and registry exported a fifth nobody imported. lib/library/compose.ts
holds the one now, with a type-only registry import so the defs files can
reach it without closing the cycle registry already opens. AUTHORING.md
tells authors to import it rather than paste it.

Dead exports, per knip. Unexported (still used inside their own file):
sharedLineStyle, TEXT_STYLES, buttonVariants, ScrollBar, anchorsOf,
withAnchor, ANCHOR_CAPTURE_PX, closestAnchor, automaticElbowAxis,
RouteAxis, ElbowHandle, CurveHandle, REVEAL_MARGIN, DistanceIndicator,
pngTargets, renderPng, renderSvg, copySelectionAsPng,
saveSelectionAsImage, copyNotice, saveNotice, CopyResult, CopyOutcome,
SaveOutcome, MIXED, controlIsVisible, isMac, ShortcutRow, ICON_WEIGHTS,
POPULAR_ICONS, basePrims, TEXT_FIRST_BASELINE, gridOf, FULL_CROP,
ContextMenuState, InspectorFocus.

Deleted outright, referenced nowhere: DropdownMenuGroup,
DropdownMenuLabel, SelectGroup, SelectLabel, SelectSeparator, searchAll,
worldToScreen, outerGroup, and snap-engine's whole distance half —
computeDistances, computePairwiseDistances, collectNodeSnapCandidates,
collectArtboardSnapCandidates, 445 lines that nothing has called since
the indicators stopped being drawn. kit stops re-exporting ICON_NAMES,
IconName, ICON_WEIGHTS and IconWeight, which nothing imported through it.

Moved, no behavior change:
- lib/sketch/paths.ts takes the DOM-free half of components/canvas/
  sketch.tsx — the rough.js generator, PEN, primOptions, roundRectPath,
  ellipsePath, drawableToPaths, primsToPaths, mirrorGlyphs,
  imagePlacement, mirrorBox. export-image no longer has to import a
  React file to print a PNG. sketch.tsx: 419 -> 179.
- components/canvas/crop-overlay.tsx and selection-overlay.tsx take the
  two overlay blocks off the end of canvas.tsx, along with the handle
  geometry the selection ring owns and crop borrows. canvas.tsx:
  3058 -> 2536.
