+++ components/ui/button.tsx
@@ -66,4 +66,4 @@ function Button({
   })
 }
 
-export { Button, buttonVariants }
+export { Button }
