+++ components/ui/scroll-area.tsx
@@ -53,4 +53,4 @@ function ScrollBar({
   )
 }
 
-export { ScrollArea, ScrollBar }
+export { ScrollArea }
